optimize code

+++ seatunnel-connectors-v2/connector-file/connector-file-ftp/src/main/java/org/apache/seatunnel/connectors/seatunnel/file/sink/ftp/util/FtpFileUtils.java
@@ -157,8 +157,11 @@ public class FtpFileUtils {
                 for (int i = 0; i < ftpFiles.length; i++) {
                     FTPFile thisFile = ftpFiles[i];
                     if (thisFile.isDirectory()) {
-                        deleteFiles(pathName + "/" + thisFile.getName());
-                        ftpClient.removeDirectory(pathName);
+                        if (deleteFiles(pathName + "/" + thisFile.getName())){
+                            ftpClient.removeDirectory(pathName);
+                        } else {
+                            return false;
+                        }
                     } else {
                         if (!ftpClient.deleteFile(pathName)) {
                             return false;
